pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["hyeonho cho <111675679+aesthetics-of-record@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
speechrecognition = "^3.11.0"
gtts = "^2.5.3"
openai = "^1.53.0"
python-multipart = "^0.0.17"
pydub = "^0.25.1"
numpy = "^2.1.2"
pyaudio = "^0.2.14"
sounddevice = "^0.5.1"
soundfile = "^0.12.1"
aiohttp = "^3.10.10"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
